requirements.txt
torch
torchvision
streamlit
Pillow
numpy
scikit-image
opencv-python
